rename llm_models
